+++ app/models/bookingsync_portal/rental.rb
@@ -5,7 +5,9 @@ class BookingsyncPortal::Rental < ActiveRecord::Base
   has_one    :connection, class_name: BookingsyncPortal.connection_model, dependent: :destroy
   has_one    :remote_rental, class_name: BookingsyncPortal.remote_rental_model, through: :connection
   has_many   :photos, class_name: BookingsyncPortal.photo_model, dependent: :destroy
-  has_many   :rates, class_name: BookingsyncPortal.rate_model, dependent: :destroy
+  if BookingsyncPortal.rate_model.present?
+    has_many   :rates, class_name: BookingsyncPortal.rate_model, dependent: :destroy
+  end
 
   validates :synced_id, uniqueness: true, presence: true